python: generalize OutputFormat class, specialize JobsOutputFormat
To generalize the OutputFormat class:

- Remove the hard-coded Jobs header names/fields
- Support replacing all number formats in headers, as opposed to just
  floating-point numbers.
- Replace "job" verbiage with generic "obj"
- Remove forced introduction of `0.` prefix to fields
- Make fields a property

To specialize the JobsOutputFormat class:

- Add back in the forced introduction of the `0.` prefix to fields via
  the `get_format` method
- Add back the hard-coded Jobs header names/fields
