+++ src/bindings/python/flux/util.py
@@ -173,43 +173,18 @@ class OutputFormat:
     a new format suitable for headers display, etc...
     """
 
-    #  List of legal format fields and their header names
-    headings = dict(
-        id="JOBID",
-        userid="UID",
-        username="USER",
-        priority="PRI",
-        state="STATE",
-        state_single="STATE",
-        name="NAME",
-        ntasks="NTASKS",
-        nnodes="NNODES",
-        nnodes_hyphen="NNODES",
-        ranks="RANKS",
-        ranks_hyphen="RANKS",
-        t_submit="T_SUBMIT",
-        t_depend="T_DEPEND",
-        t_sched="T_SCHED",
-        t_run="T_RUN",
-        t_cleanup="T_CLEANUP",
-        t_inactive="T_INACTIVE",
-        runtime="RUNTIME",
-        runtime_fsd="RUNTIME",
-        runtime_fsd_hyphen="RUNTIME",
-        runtime_hms="RUNTIME",
-    )
-
-    def __init__(self, fmt):
+    def __init__(self, valid_headings, fmt):
         """
         Parse the input format fmt with string.Formatter.
         Save off the fields and list of format tokens for later use,
         (converting None to "" in the process)
 
         Throws an exception if any format fields do not match the allowed
-        list of headings above.
+        list of headings.
         """
         from string import Formatter
 
+        self.headings = valid_headings
         self.fmt = fmt
         #  Parse format into list of (string, field, spec, conv) tuples,
         #   replacing any None values with empty string "" (this makes
@@ -219,12 +194,17 @@ class OutputFormat:
         self.format_list = [[s or "" for s in t] for t in l]
 
         #  Store list of requested fields in self.fields
-        self.fields = [field for (s, field, spec, conv) in self.format_list]
+
+        self._fields = [field for (s, field, spec, conv) in self.format_list]
 
         #  Throw an exception if any requested fields are invalid:
-        for field in self.fields:
+        for field in self._fields:
             if field and not field in self.headings:
                 raise ValueError("Unknown format field: " + field)
+
+    @property
+    def fields(self):
+        return self._fields
 
     def _fmt_tuple(self, s, field, spec, conv):
         #  If field is empty string or None, then the result of the
@@ -242,37 +222,22 @@ class OutputFormat:
         Return the header row formatted by the user-provided format spec,
         which will be made "safe" for use with string headings.
         """
-        import re
-
         l = []
         for (s, field, spec, conv) in self.format_list:
-            #  Remove floating point formatting on any spec:
-            spec = re.sub(r"\.\d+[bcdoxXeEfFgGn%]$", "", spec)
+            #  Remove number formatting on any spec:
+            spec = re.sub(r"(0?\.)?(\d+)?[bcdoxXeEfFgGn%]$", r"\2", spec)
             l.append(self._fmt_tuple(s, field, spec, conv))
         fmt = "".join(l)
         return fmt.format(**self.headings)
 
     def get_format(self):
         """
-        Return the format string with prepended `0.` if necessary.
+        Return the format string
         """
-        try:
-            return self.jobfmt
-        except AttributeError:
-            pass
+        return self.fmt
 
-        l = []
-        for (s, field, spec, conv) in self.format_list:
-            # If field doesn't have `0.` then add it
-            if field and not field.startswith("0."):
-                field = "0." + field
-            l.append(self._fmt_tuple(s, field, spec, conv))
-        self.jobfmt = "".join(l)
-        return self.jobfmt
-
-    def format(self, job):
+    def format(self, obj):
         """
-        format JobInfo object with internal format
-        prepend `0.` if necessary to fields to invoke getattr method
+        format object with internal format
         """
-        return self.get_format().format(job)
+        return self.get_format().format(obj)
