+++ src/cmd/flux-jobs.py
@@ -325,6 +325,69 @@ def parse_args():
     return parser.parse_args()
 
 
+class JobsOutputFormat(flux.util.OutputFormat):
+    """
+    Store a parsed version of the program's output format,
+    allowing the fields to iterated without modifiers, building
+    a new format suitable for headers display, etc...
+    """
+
+    #  List of legal format fields and their header names
+    headings = dict(
+        id="JOBID",
+        userid="UID",
+        username="USER",
+        priority="PRI",
+        state="STATE",
+        state_single="STATE",
+        name="NAME",
+        ntasks="NTASKS",
+        nnodes="NNODES",
+        nnodes_hyphen="NNODES",
+        ranks="RANKS",
+        ranks_hyphen="RANKS",
+        t_submit="T_SUBMIT",
+        t_depend="T_DEPEND",
+        t_sched="T_SCHED",
+        t_run="T_RUN",
+        t_cleanup="T_CLEANUP",
+        t_inactive="T_INACTIVE",
+        runtime="RUNTIME",
+        runtime_fsd="RUNTIME",
+        runtime_fsd_hyphen="RUNTIME",
+        runtime_hms="RUNTIME",
+    )
+
+    def __init__(self, fmt):
+        """
+        Parse the input format fmt with string.Formatter.
+        Save off the fields and list of format tokens for later use,
+        (converting None to "" in the process)
+
+        Throws an exception if any format fields do not match the allowed
+        list of headings above.
+        """
+        super().__init__(JobsOutputFormat.headings, fmt)
+
+    def get_format(self):
+        """
+        Return the format string with prepended `0.` if necessary.
+        """
+        try:
+            return self.jobfmt
+        except AttributeError:
+            pass
+
+        l = []
+        for (s, field, spec, conv) in self.format_list:
+            # If field doesn't have `0.` then add it
+            if field and not field.startswith("0."):
+                field = "0." + field
+            l.append(self._fmt_tuple(s, field, spec, conv))
+        self.jobfmt = "".join(l)
+        return self.jobfmt
+
+
 @flux.util.CLIMain(logger)
 def main():
     args = parse_args()
@@ -338,7 +401,7 @@ def main():
             "{ranks_hyphen}"
         )
     try:
-        of = flux.util.OutputFormat(fmt)
+        of = JobsOutputFormat(fmt)
     except ValueError as e:
         raise ValueError("Error in user format: " + str(e))
